+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/bush/OddionCrop.java
@@ -18,6 +18,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
 import net.potionstudios.biomeswevegone.world.entity.BWGEntities;
 import net.potionstudios.biomeswevegone.world.entity.oddion.Oddion;
@@ -36,12 +37,14 @@ public class OddionCrop extends BWGBerryBush {
 
     @Override
     public @NotNull InteractionResult use(BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand hand, @NotNull BlockHitResult hit) {
-        if (player.getItemInHand(hand).getItem() == Items.BONE_MEAL)
+        if (player.getItemInHand(hand).is(Items.BONE_MEAL))
             if (state.getValue(AGE) == MAX_AGE && !state.getValue(HATCHING)) {
-                level.setBlockAndUpdate(pos, state.cycle(HATCHING));
+                BlockState blockState = state.cycle(HATCHING);
+                level.setBlockAndUpdate(pos, blockState);
+                level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, state));
                 level.addParticle(ParticleTypes.HAPPY_VILLAGER, (double)pos.getX() + 0.5, (double)pos.getY() + 0.5, (double)pos.getZ() + 0.5, 0.0, 0.0, 0.0);
                 level.playLocalSound(pos, SoundEvents.ENCHANTMENT_TABLE_USE, SoundSource.AMBIENT, 2 ,1, false);
-                return InteractionResult.SUCCESS;
+                return InteractionResult.sidedSuccess(level.isClientSide());
             }
         return InteractionResult.PASS;
     }